==> 03_2_FIAS_clean_files.ipynb <==
Удаляем лишние из датанные из датасетов.

В ФИАС есть некоторое количество некорректных значений, поэтому иногда придется выгружать датасеты в Эксель и править вручную.

import pandas as pd
import numpy as np

pd.set_option('precision', 13)
pd.options.display.max_rows = 100

# загружаем города и поселки

all_cities = pd.read_csv('C:/00_Projects/Lenin/01_FIAS_csv_post/03_all_cities.csv')

lenin_cities = pd.read_csv('C:/00_Projects/Lenin/01_FIAS_csv_post/04_lenin_cities.csv')

# загружаем улицы

all_streets = pd.read_csv('C:/00_Projects/Lenin/01_FIAS_csv_post/05_all_streets.csv')

lenin_streets = pd.read_csv('C:/00_Projects/Lenin/01_FIAS_csv_post/06_lenin_streets.csv')

lenin_streets

## Чистим города и поселки ##

Чистим all_cities и lenin_cities

all_cities.isnull().sum()

all_cities.shape

all_cities

# смотрим сколько вообще  типов населенных пунктов есть в базе

all_cities.city_type.nunique()

# и каких

(all_cities
.groupby('city_type', as_index=False)
.agg({'aolevel': 'count'})
.rename(columns={'aolevel': 'city_number'})
.sort_values('city_number', ascending=False)
)

# группируем населенные пункты, где есть улицы

city_type_with_streets = (all_streets
 .groupby(['region', 'mun_district', 'city_type', 'city'], as_index=False)
 .agg({'street': 'count'})
 .rename(columns={'street': 'street_number'})                          
 .sort_values('street_number', ascending=False)
                          
)

# количество типов пунктов с улицами

city_type_with_streets.city_type.nunique()

city_type_with_streets.city_type.value_counts().reset_index()

Посмотрим ленинские города и города с ленинскими улицами

# типы "ленинских городов"

lenin_cities.city_type.nunique()

# типы городов и поселков, где есть ленинские улицы

lenin_streets.city_type.nunique()

# ленинские города и поселки

lenin_city_type = (lenin_streets
                   .groupby(['region', 'mun_district', 'city_type', 'city'], as_index=False)
                   .agg({'street': 'count'})
                   .rename(columns={'street': 'street_number'})                          
                   .sort_values('street_number', ascending=False)
                  )

lenin_city_type.city_type.value_counts().reset_index()

Не будем искать Ленина в кустах (хотя он там есть) и удалим часть населенных пунктов из базы.

Оставлены аалы, арбаны, выселы, починки и пр. (это, по сути, названия сел).

Убираем 'р-н', 'сп', 'сп.' -- это улицы без населенного пункта, "повисшие" в районе или сельском поселении.

# выгружаем все типы

print(all_cities.city_type.value_counts().index.to_list())

no_streets = ['автодорога',
              'ж/д б-ка', 'ж/д пл-ка', 'ж/д_будка',
              'ж/д_казарм', 'ж/д_оп', 'ж/д_платф', 'ж/д_пост', 'ж/д_рзд',
              'заимка', 'казарма', 'кордон', 'лпх',
              'м', 'п/о', 'погост', 'пос.рзд', 'промзона', 'рзд',
              'с/мо', 'с/о', 'с/с', 'снт', 'тер',
              'р-н']

### Удаляем ###

**Все города**

all_cities.shape

# Удаляем ненужные записи

all_cities = all_cities.loc[~all_cities.city_type.isin(no_streets)]

all_cities.shape

**Ленинские города**

lenin_cities.shape

# Удаляем ненужные записи

lenin_cities = lenin_cities.loc[~lenin_cities.city_type.isin(no_streets)]

lenin_cities.shape

**Все улицы**

all_streets.shape

# Удаляем ненужные записи

all_streets = all_streets.loc[~all_streets.city_type.isin(no_streets)]

all_streets.shape

**Ленинские улицы**

lenin_streets.shape

# Удаляем ненужные записи

lenin_streets = lenin_streets.loc[~lenin_streets.city_type.isin(no_streets)]

lenin_streets.shape

## Чистим улицы ##

### Удаляем ненужные записи ###

Нам не нужны отдельные населенные пункты (без улиц), промышленные объекты, территории, садовые товарищества, парки, дороги и т.п.

# смотрим, что вообще там есть (это уровень 7)

all_streets.shortname.value_counts().index.to_list

# отбираем ненужные в список

unnecessary_list = (['а/я', 'берег', 'г-к', 'гск', 'д', 'дор',
  'ж/д_будка', 'ж/д_казарм', 'ж/д_оп', 'ж/д_платф', 'ж/д_пост', 'ж/д_рзд', 'ж/д_ст',
  'жт', 'зона', 'казарма', 'кв-л', 'км', 'коса', 'м', 'мгстр.', 'местность', 'мкр',
  'мост', 'нп', 'остров', 'п/о', 'п/ст', 'парк', 'пер-д', 'переезд', 'платф', 'промзона',
  'рзд.', 'с', 'с-р', 'с/т', 'сад', 'сл', 'снт', 'ст', 'стр', 'тер', 'тер. СНТ', 'уч-к',
  'ф/х', 'ферма', 'х'])

Полный список с расшифровкой (удаляемого):

- а/я - абонентский ящик
- берег - берег
- г-к - городок
- гск - гаражно-строительный кооператив
- д - деревня
- дор - дорога
- ж/д_будка - жд будка
- ж/д_казарм - жд казарма
- ж/д_оп - жд отдел полиции
- ж/д_платф - жд платформа
- ж/д_пост - жд пост
- ж/д_рзд - жд разъезд
- ж/д_ст - жд станция
- жт - животноводческая точка
- зона - зона
- казарма - казарма
- кв-л - квартал
- км - километр
- коса - коса
- м - местечко
- мгстр. - магистраль
- местность - местность
- мкр - микрарайон
- мост - мост
- нп - населенный пункт
- остров - остров
- п/о - почтовое отделение
- п/ст - поселок при станции
- парк - парк
- пер-д - переезд
- переезд - переезд
- платф - платформа
- промзона - промзона
- рзд. - разъезд
- с - село
- с-р - сквер
- с/т - садовое товарищество
- сад - сад
- сл - слобода
- снт - садовое некоммерческое товарищество
- ст - станция
- стр - строение
- тер - территория
- тер. СНТ - территория садового некоммерческого товарищества
- уч-к - участок
- ф/х - фермерское хозяйство
- ферма - ферма
- х - хутор

all_streets.shape

# Удаляем ненужные записи из all_streets

all_streets = all_streets.loc[~all_streets.shortname.isin(unnecessary_list)]

all_streets.shape

# смотрим, что осталось

all_streets.shortname.value_counts().index.to_list

lenin_streets.shape

# Удаляем ненужные записи из lenin_streets

lenin_streets = lenin_streets.loc[~lenin_streets.shortname.isin(unnecessary_list)]

lenin_streets.shape

Удаляем улицы, в названии которых есть "км.", "км", "автодорога", "автомагистраль" (это, можно сказать, технические улицы, участки автотрасс).

extra_streets = [' км', 'километр', 'автодорог', 'автомагистраль ', ' массив']

# удаляем лишние записи по названию (через цикл, хотя это не лучший метод, но работает).

for i in extra_streets:
    all_streets = all_streets.drop(all_streets[all_streets.street.str.contains(i, case=False)].index)

all_streets.shape

# удаляем лишние записи в ленинских

for i in extra_streets:
    lenin_streets = lenin_streets.drop(lenin_streets[lenin_streets.street.str.contains(i, case=False)].index)

lenin_streets.shape

all_streets = all_streets.drop('postalcode', axis=1)

all_streets.shape

all_streets = all_streets.drop_duplicates()

all_streets.shape

## Записываем файлы в Excel ##

К сожалению, на автомате из-за ошибок и поправок в ФИАС все сделать не получится.

all_cities.to_excel('C:/00_Projects/Lenin/01_FIAS_excel/03_all_cities.xlsx', index=False)

lenin_cities.to_excel('C:/00_Projects/Lenin/01_FIAS_excel/04_lenin_cities.xlsx', index=False)

# all_streets в csv (в Экселе с таким объемом трудно работать)

all_streets.to_csv('C:/00_Projects/Lenin/01_FIAS_excel/05_all_streets.csv', sep=',', encoding='utf-8', index=False)

lenin_streets.to_excel('C:/00_Projects/Lenin/01_FIAS_excel/06_lenin_streets.xlsx', index=False)
