==> 04_1_FIAS_region_preprocessing.ipynb <==
import pandas as pd
import numpy as np


import matplotlib.pyplot as plt
import seaborn as sns

%matplotlib inline

## Загружаем датасеты ##

Из исправленных и очищенных файлов

### Города и поселки ###

# загружаем все города и поселки

all_cities = pd.read_excel('C:/00_Projects/Lenin/Datasets/01_FIAS/03_all_cities.xlsx', sheet_name='cities')

# загружаем ленинские города и поселки, только строго "ленинские"

lenin_cities = pd.read_excel('C:/00_Projects/Lenin/Datasets/01_FIAS/04_lenin_cities.xlsx', sheet_name='lenin_strict')

### Улицы ###

# загружаем все улицы

work_streets = pd.read_csv('C:/00_Projects/Lenin/Datasets/01_FIAS/05_all_streets.csv')

# загружаем ленинские улицы (полный вариант)

lenin_streets_full = pd.read_excel('C:/00_Projects/Lenin/Datasets/01_FIAS/06_lenin_streets.xlsx', sheet_name='full')

# загружаем ленинские улицы (только точные)

lenin_streets_strict = pd.read_excel('C:/00_Projects/Lenin/Datasets/01_FIAS/06_lenin_streets.xlsx', sheet_name='lenin')

# загружаем ленинские улицы (без переулков)

lenin_only_streets = pd.read_excel('C:/00_Projects/Lenin/Datasets/01_FIAS/06_lenin_streets.xlsx', sheet_name='lenin_os')

### Памятники ###

# Загружаем памятники из ОКН

monuments_OKN = pd.read_excel('C:/00_Projects/Lenin/Datasets/02_Monuments/08_lenin_reestrOKN.xlsx',
                                     sheet_name='need')

# Загружаем памятники из базы leninstatues

monuments_lst = pd.read_excel('C:/00_Projects/Lenin/Datasets/02_Monuments/09_lenin_monuments.xlsx',
                                     sheet_name='need')

### Названия регионов для Datawrapper ###

#Загружаем названия регионов для datawrapper

region_name = pd.read_excel('C:/00_Projects/Lenin/Datasets/03_Other/region_name.xlsx')

# создаем рабочий датафрейм

region_work = region_name[['datawrapperName', 'region']]

## Считаем и группируем ##

### Города и поселки ###

# считаем количество населенных пунктов в регионах

all_cities_temp = (all_cities
                   .groupby('region', as_index=False)
                   .agg({'city': 'count'})
                   .rename(columns={'city': 'total_cities'})
                  )

# считаем "ленинские" населенные пункты в регионах

lenin_cities_temp = (lenin_cities
                     .groupby('region', as_index=False)
                     .agg({'city': 'count'})
                     .rename(columns={'city': 'lenin_cities'})
                    )

# считаем типы ленинских населенных пунктов

(lenin_cities
 .groupby('city_type', as_index=False)
 .agg({'city': 'count'})
 .rename(columns={'city': 'number'})
 .sort_values('number', ascending=False)
)

#.to_excel('C:/00_Projects/Lenin/Datasets/00_for_viz/temp.xlsx', index=False)

### Считаем населенные пункты с количеством улиц больше 5 и 10 ###

# Количество объектов (населенных пунктов) в ФИАС

all_cities.shape

# считаем количество улиц по населенным пунктам

city_street_number = (work_streets
                      .groupby(['region', 'mun_district', 'city', 'city_type'], as_index=False)
                      .agg({'street': 'count'})
                      .rename(columns={'street': 'street_number'})
       )

# количество насленных пунктов с улицами

city_street_number.shape

# количество объектов без улиц

all_cities.shape[0] - city_street_number.shape[0]

# количество объектов без улиц (вычитаем из 165 тыс. - числа объектов в ФИАС до чистки)

165038 - city_street_number.shape[0]

# считаем среднее количество улиц в населенных пунктах с улицами

city_street_number.street_number.mean()

# считаем медиану количества улиц в населенных пунктах с улицами

city_street_number.street_number.median()

city_street_number.describe()

city_street_number.head(5)

# города и поселки с 5 и более улицами

city_more_5_streets = city_street_number.loc[city_street_number.street_number >= 5]

# их количество

city_more_5_streets.shape

# считаем количество населенных пунктов с 5 и более улицами

city_more_5_streets_temp = (city_more_5_streets
                            .groupby('region', as_index=False)
                            .agg({'city': 'count'})
                            .rename(columns={'city': 'city_m5sts'})
                            .sort_values('city_m5sts', ascending=False)
                           )
 

city_more_5_streets_temp.head(10)

# города и поселки с 10 и более улицами

city_more_10_streets = city_street_number.loc[city_street_number.street_number >= 10]

# их количество

city_more_10_streets.shape

# считаем количество населенных пунктов с 10 и более улицами

city_more_10_streets_temp = (city_more_10_streets
                            .groupby('region', as_index=False)
                            .agg({'city': 'count'})
                            .rename(columns={'city': 'city_m10sts'})
                            .sort_values('city_m10sts', ascending=False)
                           )
 

city_more_10_streets_temp.head(10)

### Считаем сумму ленинских улиц в городах с количеством улиц больше 5 и 10 ###

**Населенные пункты с количеством улиц больше 5**

# группируем города и поселки с ленинскими улицами

city_street_number_lsts = (lenin_streets_strict
                           .groupby(['region', 'mun_district', 'city', 'city_type'], as_index=False)
                           .agg({'street': 'count'})
                           .rename(columns={'street': 'lenin_street_number'})
                          )

# считаем их количество

city_street_number_lsts.shape

city_street_number_lsts.head()

# объединяем с city_more_5_streets, чтобы отсеять нас.пункты с количеством улиц меньше 5

city_more_5_street_lsts = city_street_number_lsts.merge(
    city_more_5_streets,
    on=['region', 'mun_district', 'city', 'city_type'], how='left')

# заполняем пропуски

city_more_5_street_lsts.fillna(0, inplace=True)

city_more_5_street_lsts.shape

# убираем нас.пункты с количеством улиц меньше 5

city_more_5_street_lsts = city_more_5_street_lsts[city_more_5_street_lsts.street_number > 0]

city_more_5_street_lsts.shape

# считаем сумму улиц в этих городах и поселках по регионам

lenin_street_sum_cm5sts_temp = (city_more_5_street_lsts
                                .groupby('region', as_index=False)
                                .agg({'lenin_street_number': 'sum'})
                                .rename(columns={'lenin_street_number': 'lenin_street_cm5st'})
                                .sort_values('lenin_street_cm5st', ascending=False)
                               )

lenin_street_sum_cm5sts_temp.head(10)

**Населенные пункты с количеством улиц больше 10**

# объединяем с city_more_10_street, чтобы отсеять нас.пункты с количеством улиц меньше 5

city_more_10_street_lsts = city_street_number_lsts.merge(
    city_more_10_streets,
    on=['region', 'mun_district', 'city', 'city_type'], how='left')

# заполняем пропуски

city_more_10_street_lsts.fillna(0, inplace=True)

city_more_10_street_lsts.shape

# убираем нас.пункты с количеством улиц меньше 10

city_more_10_street_lsts = city_more_10_street_lsts[city_more_10_street_lsts.street_number > 0]

city_more_10_street_lsts.shape

# считаем сумму улиц в этих городах и поселках по регионам

lenin_street_sum_cm10sts_temp = (city_more_10_street_lsts
                                 .groupby('region', as_index=False)
                                 .agg({'lenin_street_number': 'sum'})
                                 .rename(columns={'lenin_street_number': 'lenin_street_cm10st'})
                                 .sort_values('lenin_street_cm10st', ascending=False)
                                )

lenin_street_sum_cm10sts_temp.head(10)

### Количество городов и поселков, в которых есть улицы Ленина ###

# считаем количество всех городов, в которых есть хотя бы одна улица Ленина

all_cities_lenin_streets_temp = (lenin_streets_strict
                                 .drop_duplicates(['city', 'mun_district', 'region'])
                                 .groupby('region', as_index=False)
                                 .agg({'city': 'count'})
                                 .rename(columns={'city': 'cities_1_lsts'})
                                )

Cчитаем количество городов, в которых есть несколько улиц Ленина.

# берем датасет lenin_only_streets, чтобы не считать переулки и группируем города по количеству улиц

all_cities_lenin_only_streets = (lenin_only_streets
                                 .groupby(['region', 'mun_district', 'city', 'city_type'], as_index=False)
                                 .agg({'street': 'count'})
                                 .rename(columns={'street': 'lenin_street_number'})
                                 .sort_values('lenin_street_number', ascending=False)
                                )

# смотрим среднее значение

all_cities_lenin_only_streets.lenin_street_number.mean()

# считаем количество городов, в которых 2 и больше улицы Ленина

all_cities_2_lenin_streets_temp = (all_cities_lenin_only_streets
                                   [all_cities_lenin_only_streets.lenin_street_number >= 2]
                                   .groupby('region', as_index=False)
                                   .agg({'city': 'count'})
                                   .rename(columns={'city': 'cities_2_lsts'})
                                   .sort_values('cities_2_lsts', ascending=False)
                                  )

### Города и поселки с улицами Ленина с общим количество улиц больше 5 и 10 ###

**Города и поселки с количеством улиц больше 5**

# считаем количество населенных пунктов, где есть хотя бы одна улица Ленина, с общим количеством улиц больше 5

cities_m5st_1_lst_temp = (city_more_5_street_lsts
                                .groupby('region', as_index=False)
                                .agg({'lenin_street_number': 'count'})
                                .rename(columns={'lenin_street_number': 'cities_1_lsts_cm5s'})
                                .sort_values('cities_1_lsts_cm5s', ascending=False)
                               )

cities_m5st_1_lst_temp.head(10)

# считаем количество населенных пунктов, где есть как минимум 2 улицы Ленина, с общим количеством улиц больше 5
# берем все города с количеством улиц больше 5
# и присоединяем к ним города с ленискими улицами в варианте без переулков - all_cities_lenin_only_streets


cities_lenin_only_streets_cm5sts = city_more_5_streets.merge(all_cities_lenin_only_streets,
                          on=(['region', 'mun_district', 'city', 'city_type']),
                              how='left'
                             )

# заполняем пустые значения

cities_lenin_only_streets_cm5sts.fillna(0, inplace=True)

cities_lenin_only_streets_cm5sts.shape

# удаляем нулевые строки, т.е. нас.пункты, где нет ленинских улиц

cities_lenin_only_streets_cm5sts = (cities_lenin_only_streets_cm5sts
                                    .loc[cities_lenin_only_streets_cm5sts.lenin_street_number > 0]
                                   )

cities_lenin_only_streets_cm5sts.shape

# считаем количество нас.пунктов с количеством ленинских улиц 2 и больше

cities_m5st_2_lst_temp = (cities_lenin_only_streets_cm5sts
                          .loc[cities_lenin_only_streets_cm5sts.lenin_street_number >= 2]
                          .groupby('region', as_index=False)
                          .agg({'city': 'count'})
                          .rename(columns={'city': 'cities_2_lsts_cm5s'})
                          .sort_values('cities_2_lsts_cm5s', ascending=False)
                         )

cities_m5st_2_lst_temp.head(10)

**Города и поселки с количеством улиц больше 10**

# считаем количество населенных пунктов, где есть хотя бы одна улица Ленина, с общим количеством улиц больше 10

cities_m10st_1_lst_temp = (city_more_10_street_lsts
                                .groupby('region', as_index=False)
                                .agg({'lenin_street_number': 'count'})
                                .rename(columns={'lenin_street_number': 'cities_1_lsts_cm10s'})
                                .sort_values('cities_1_lsts_cm10s', ascending=False)
                               )

cities_m10st_1_lst_temp.head(10)

# считаем количество населенных пунктов, где есть как минимум 2 улицы Ленина, с общим количеством улиц больше 10
# повторяем предыдущий код
# берем все города с количеством улиц больше 10
# и присоединяем к ним города с ленискими улицами в варианте без переулков - all_cities_lenin_only_streets


cities_lenin_only_streets_cm10sts = city_more_10_streets.merge(all_cities_lenin_only_streets,
                          on=(['region', 'mun_district', 'city', 'city_type']),
                              how='left'
                             )

# заполняем пустые значения

cities_lenin_only_streets_cm10sts.fillna(0, inplace=True)

cities_lenin_only_streets_cm10sts.shape

# удаляем нулевые строки, т.е. нас.пункты, где нет ленинских улиц

cities_lenin_only_streets_cm10sts = (cities_lenin_only_streets_cm10sts
                                    .loc[cities_lenin_only_streets_cm10sts.lenin_street_number > 0]
                                   )

cities_lenin_only_streets_cm10sts.shape

# считаем количество нас.пунктов с количеством ленинских улиц 2 и больше

cities_m10st_2_lst_temp = (cities_lenin_only_streets_cm10sts
                          .loc[cities_lenin_only_streets_cm10sts.lenin_street_number >= 2]
                          .groupby('region', as_index=False)
                          .agg({'city': 'count'})
                          .rename(columns={'city': 'cities_2_lsts_cm10s'})
                          .sort_values('cities_2_lsts_cm10s', ascending=False)
                         )

cities_m10st_2_lst_temp.head(10)

### Количество улиц ###

# считаем все улицы

all_streets_temp = (work_streets
                    .groupby('region', as_index=False)
                    .agg({'street': 'count'})
                    .rename(columns={'street': 'total_streets'})
                   )

# считаем ленинские улицы (полный вариант)

lenin_streets_full_temp = (lenin_streets_full
                           .groupby('region', as_index=False)
                           .agg({'street': 'count'})
                           .rename(columns={'street': 'len_streets_full'})
                          )

# считаем ленинские улицы (строгий вариант)

lenin_streets_strict_temp = (lenin_streets_strict
                             .groupby('region', as_index=False)
                             .agg({'street': 'count'})
                             .rename(columns={'street': 'len_streets_strict'})
                            )

# считаем главные улицы (оставляем только улицы, площади, проспекты, бульвары, шоссе)

lenin_only_streets_temp = (lenin_only_streets
                           .groupby('region', as_index=False)
                           .agg({'street': 'count'})
                           .rename(columns={'street': 'len_only_streets'})
                          )

### Памятники ###

# считаем памятники из ОКН

monuments_OKN_temp = (monuments_OKN
                      .groupby('region', as_index=False)
                      .agg({'monument': 'count'})
                      .rename(columns={'monument': 'monuments_OKN'})
                     )

# считаем памятники из базы lst

monuments_lst_temp = (monuments_lst
                      .groupby('region', as_index=False)
                      .agg({'original_name': 'count'})
                      .rename(columns={'original_name': 'monuments_lst'})
                     )

### Считаем количество памятников в населенных пунктах с количеством улиц больше 5 и 10 ###

**Больше 5 улиц в нас.пункте ОКН**

# группируем города и поселки с монументами из ОКН

cities_monuments_OKN = (monuments_OKN
                        .groupby(['region', 'mun_district', 'city', 'city_type'], as_index=False)
                        .agg({'monument': 'count'})
                        .rename(columns={'monument': 'monument_OKN_number'})
                       )

# объединяем сгруппированные нас.пункты с city_w_more_5_street, чтобы отсеять нас.пункты с количеством улиц меньше 5

city_more_5_streets_OKN = cities_monuments_OKN.merge(
    city_more_5_streets,
    on=['region', 'mun_district', 'city', 'city_type'], how='left')

# заполняем пропуски

city_more_5_streets_OKN.fillna(0, inplace=True)

city_more_5_streets_OKN.shape

# удаляем нас.пункты с количеством улиц меньше 5

city_more_5_streets_OKN = city_more_5_streets_OKN[city_more_5_streets_OKN.street_number > 0]

city_more_5_streets_OKN.shape

# считаем количество памятников по регионам

monument_OKN_cm5sts_temp = (city_more_5_streets_OKN
                              .groupby('region', as_index=False)
                              .agg({'monument_OKN_number': 'count'})
                              .rename(columns={'monument_OKN_number': 'monument_OKN_cm5sts'})
                              .sort_values('monument_OKN_cm5sts', ascending=False)
                             )

monument_OKN_cm5sts_temp.head(10)

**Больше 10 улиц в нас.пункте ОКН**

# объединяем сгруппированные нас.пункты с city_w_more_5_street, чтобы отсеять нас.пункты с количеством улиц меньше 5

city_more_10_streets_OKN = cities_monuments_OKN.merge(
    city_more_10_streets,
    on=['region', 'mun_district', 'city', 'city_type'], how='left')

# заполняем пропуски

city_more_10_streets_OKN.fillna(0, inplace=True)

city_more_10_streets_OKN.shape

# удаляем нас.пункты с количеством улиц меньше 10

city_more_10_streets_OKN = city_more_10_streets_OKN[city_more_10_streets_OKN.street_number > 0]

city_more_10_streets_OKN.shape

# считаем количество памятников по регионам

monument_OKN_cm10sts_temp = (city_more_10_streets_OKN
                               .groupby('region', as_index=False)
                               .agg({'monument_OKN_number': 'count'})
                               .rename(columns={'monument_OKN_number': 'monument_OKN_cm10sts'})
                               .sort_values('monument_OKN_cm10sts', ascending=False)
                              )

monument_OKN_cm10sts_temp.head(10)

**Больше 5 улиц в нас.пункте leninstatues**

# группируем города и поселки с монументами из leninstatues

cities_monuments_lst = (monuments_lst
                          .groupby(['region', 'mun_district', 'city', 'city_type'], as_index=False)
                          .agg({'status': 'count'})
                          .rename(columns={'status': 'monument_lst_number'})
                          .sort_values('monument_lst_number', ascending=False)
       )

# объединяем сгруппированные нас.пункты с city_more_5_streets, чтобы отсеять нас.пункты с количеством улиц меньше 5

city_more_5_streets_lst = cities_monuments_lst.merge(
    city_more_5_streets,
    on=['region', 'mun_district', 'city', 'city_type'], how='left')

# заполняем пропуски

city_more_5_streets_lst.fillna(0, inplace=True)

city_more_5_streets_lst.shape

# удаляем нас.пункты с количеством улиц меньше 5

city_more_5_streets_lst = city_more_5_streets_lst[city_more_5_streets_lst.street_number > 0]

city_more_5_streets_lst.shape

# считаем количество памятников по регионам

monument_lst_cm5sts_temp = (city_more_5_streets_lst
                            .groupby('region', as_index=False)
                            .agg({'monument_lst_number': 'count'})
                            .rename(columns={'monument_lst_number': 'monument_lst_cm5sts'})
                            .sort_values('monument_lst_cm5sts', ascending=False)
                           )

monument_lst_cm5sts_temp.head(10)

**Больше 10 улиц в нас.пункте leninstatues**

# объединяем сгруппированные нас.пункты с city_more_10_street, чтобы отсеять нас.пункты с количеством улиц меньше 10

city_more_10_streets_lst = cities_monuments_lst.merge(
    city_more_10_streets,
    on=['region', 'mun_district', 'city', 'city_type'], how='left')

# заполняем пропуски

city_more_10_streets_lst.fillna(0, inplace=True)

city_more_10_streets_lst.shape

# удаляем нас.пункты с количеством улиц меньше 10

city_more_10_streets_lst = city_more_10_streets_lst[city_more_10_streets_lst.street_number > 0]

city_more_10_streets_lst.shape

# считаем количество памятников по регионам

monument_lst_cm10sts_temp = (city_more_10_streets_lst
                            .groupby('region', as_index=False)
                            .agg({'monument_lst_number': 'count'})
                            .rename(columns={'monument_lst_number': 'monument_lst_cm10sts'})
                            .sort_values('monument_lst_cm10sts', ascending=False)
                           )

monument_lst_cm10sts_temp.head(10)

## Объединяем датафреймы ##

# присоединяем к рабочему датафрейму количество ленинских городов и поселков (названных в честь Ильича)

region_work = region_work.merge(lenin_cities_temp, on='region', how='left')

# присоединяем все города и поселки

region_work = region_work.merge(all_cities_temp, on='region', how='left')

# присоединяем города с количеством улиц 5 и больше и 10 и больше

region_work = region_work.merge(city_more_5_streets_temp, on='region', how='left')

region_work = region_work.merge(city_more_10_streets_temp, on='region', how='left')

# присоединяем количество городов, где есть улицы Ленина

# хотя бы одна улица Ленина
region_work = region_work.merge(all_cities_lenin_streets_temp, on='region', how='left')

# как минимум, 2
region_work = region_work.merge(all_cities_2_lenin_streets_temp, on='region', how='left')

# присоединяем нас.пункты с количеством улиц больше 5 и 10, где есть улицы Ленина

# 
region_work = region_work.merge(cities_m5st_1_lst_temp, on='region', how='left')

region_work = region_work.merge(cities_m5st_2_lst_temp, on='region', how='left')


# присоединяем города с количеством улиц больше 5 и 10, где есть улицы Ленина

region_work = region_work.merge(cities_m10st_1_lst_temp, on='region', how='left')

region_work = region_work.merge(cities_m10st_2_lst_temp, on='region', how='left')

# присоединяем улицы

region_work = region_work.merge(all_streets_temp, on='region', how='left')


region_work = region_work.merge(lenin_streets_full_temp, on='region', how='left')


region_work = region_work.merge(lenin_streets_strict_temp, on='region', how='left')


region_work = region_work.merge(lenin_only_streets_temp, on='region', how='left')

# присоединяем города с количеством улиц больше 5 и 10, где есть улицы Ленина

region_work = region_work.merge(lenin_street_sum_cm5sts_temp, on='region', how='left')

region_work = region_work.merge(lenin_street_sum_cm10sts_temp, on='region', how='left')

# присоединяем памятники

region_work = region_work.merge(monuments_OKN_temp, on='region', how='left')

region_work = region_work.merge(monuments_lst_temp, on='region', how='left')

# присоединяем памятники в городах с количеством улиц больше 5 и 10

region_work = region_work.merge(monument_OKN_cm5sts_temp, on='region', how='left')

region_work = region_work.merge(monument_lst_cm5sts_temp, on='region', how='left')

region_work = region_work.merge(monument_OKN_cm10sts_temp, on='region', how='left')

region_work = region_work.merge(monument_lst_cm10sts_temp, on='region', how='left')

# заполняем Nan-значения

region_work.fillna(0, inplace=True)

region_work

## Записываем результат ##

# записываем csv

region_work.to_csv('C:/00_Projects/Lenin/Datasets/04_work_data/work_region.csv', sep=',', encoding='utf-8', index=False)
