==> 08_for_vz.ipynb <==
import pandas as pd
import numpy as np

import re


import matplotlib.pyplot as plt
import seaborn as sns

%matplotlib inline

### Города и поселки ###

# загружаем все города и поселки

all_cities = pd.read_excel('C:/00_Projects/Lenin/Datasets/01_FIAS/03_all_cities.xlsx', sheet_name='cities')

### Улицы ###

# загружаем все улицы

work_streets = pd.read_csv('C:/00_Projects/Lenin/Datasets/01_FIAS/05_all_streets.csv')

# загружаем ленинские улицы (включая косвенные)

lenin_streets_full = pd.read_excel('C:/00_Projects/Lenin/Datasets/01_FIAS/06_lenin_streets.xlsx', sheet_name='full')

# загружаем ленинские улицы (только точные)

lenin_streets_strict = pd.read_excel('C:/00_Projects/Lenin/Datasets/01_FIAS/06_lenin_streets.xlsx', sheet_name='lenin')

# загружаем ленинские улицы (без переулков)

lenin_only_streets = pd.read_excel('C:/00_Projects/Lenin/Datasets/01_FIAS/06_lenin_streets.xlsx', sheet_name='lenin_os')

### Памятники ###

# Загружаем памятники из ОКН

lenin_monuments_OKN = pd.read_excel('C:/00_Projects/Lenin/Datasets/02_Monuments/08_lenin_reestrOKN.xlsx',
                                     sheet_name='need')

# Загружаем памятники из базы leninstatues

lenin_monuments_lst = pd.read_excel('C:/00_Projects/Lenin/Datasets/02_Monuments/09_lenin_monuments.xlsx',
                                     sheet_name='need')

### Столицы регионов ###

# загружаем столицы регионов

reg_capitals = pd.read_excel('C:/00_Projects/Lenin/Datasets/03_Other/region_name.xlsx', sheet_name='reg_capitals')

### Города с населением больше 100 тыс. ###

cities_100 = pd.read_excel('C:/00_Projects/Lenin/Datasets/03_Other/cities_rf_100_ph.xlsx', sheet_name='cities_100')

### Названия регионов для Datawrapper ###

region_name = pd.read_excel('C:/00_Projects/Lenin/Datasets/03_Other/region_name.xlsx', sheet_name='reg_name')

region_work_only_city = region_name[['datawrapperName', 'region']]

### Viz №1, улицы Ленина в топ-100 ###

top_100_streets = (work_streets
 .groupby('street', as_index=False)
 .agg({'aolevel': 'count'})
 .rename(columns={'aolevel': 'street_number'})
 .sort_values('street_number', ascending=False)
                  ).head(100) 

top_100_streets.head(15)

### Viz №2-3, проспекты и площади ###

# проспекты

top_15_avenues = (work_streets[work_streets.shortname == 'пр-кт']
                  .groupby('street', as_index=False)
                  .agg({'aolevel': 'count'})
                  .rename(columns={'street': 'avenue', 'aolevel': 'avenue_number'})
                  .sort_values('avenue_number', ascending=False)
                 )

top_15_avenues.head()

# площади

top_15_squeres = (work_streets[work_streets.shortname == 'пл']
                  .groupby('street', as_index=False)
                  .agg({'aolevel': 'count'})
                  .rename(columns={'street': 'squere', 'aolevel': 'squere_number'})
                  .sort_values('squere_number', ascending=False)
                 )

top_15_squeres.head(15)

### 05 Таблица: типы ленинских улиц  ###

(lenin_streets_strict
 .shortname.value_counts()
 .reset_index()
 .rename(columns={'index': 'street_type', 'shortname': 'number'})
)

### 06 Таблица: всех ленинских улиц  ###

lenin_streets_full.shape

lenin_streets_strict.shape

lenin_only_streets.shape

### 07 Количество ленинских улиц по регионам ###

len_streets_for_map = (lenin_streets_strict
                       .groupby('region', as_index=False)
                       .agg({'street': 'count'})
                       .rename(columns={'street': 'len_street_strict'})
                      )

len_streets_full_for_map = (lenin_streets_full
                       .groupby('region', as_index=False)
                       .agg({'street': 'count'})
                       .rename(columns={'street': 'len_street_full'})
                      )

len_streets_for_map = len_streets_for_map.merge(len_streets_full_for_map, on='region', how='left')

len_streets_for_map.head()

## Памятники ##

# группируем по годам и регионам

add_year = (lenin_monuments_OKN
 .groupby(['add_year', 'region'], as_index=False)
 .agg({'monument': 'count'})
 .rename(columns={'add_year': 'year', 'monument': 'number'})
 .sort_values(['year', 'number'], ascending=[True, False])
)

add_year

### 11 Количество памятников по регионам ###

# считаем памятники по регионам

all_monuments_for_map = (lenin_monuments_lst
                        .groupby('region', as_index=False)
                        .agg({'original_name': 'count'})
                        .rename(columns={'original_name': 'monument_number'})
                       )

# считаем памятники ОКН по регионам

OKN_monuments_for_map = (lenin_monuments_OKN
                        .groupby('region', as_index=False)
                        .agg({'monument': 'count'})
                        .rename(columns={'monument': 'OKN_number'})
                       )

all_monuments_for_map = all_monuments_for_map.merge(OKN_monuments_for_map, on='region', how='left')

all_monuments_for_map.fillna(0, inplace=True)

all_monuments_for_map

#.to_excel('C:/00_Projects/Lenin/Datasets/05_for_viz/temp.xlsx', encoding='utf-8', index=False)

### Есть ли улица Ленина в каждом городе? ###

### Процент населеных пунктов, где есть ленинские улицы ###

cities_with_len_st = (lenin_streets_strict
                     .groupby(['region', 'mun_district', 'city_type', 'city'], as_index=False)
                     .agg({'street': 'count'})
                     .rename(columns={'street': 'len_street_number'})
                    )

cities_with_mon = (lenin_monuments_lst
                     .groupby(['region', 'mun_district', 'city_type', 'city'], as_index=False)
                     .agg({'original_name': 'count'})
                     .rename(columns={'original_name': 'mon_number'})
                    )

all_cities_with_streets = (work_streets
                           .groupby(['region', 'mun_district', 'city_type', 'city'], as_index=False)
                           .agg({'street': 'count'})
                           .rename(columns={'street': 'total_street'})
                          )

cities_with_len_st.shape[0] / all_cities_with_streets.shape[0]

cities_with_mon.shape[0] / all_cities_with_streets.shape[0]

cities_with_lenin

### Города с переулками Ленина, но без улиц ###

# агрегируем все улицы

cities_with_len_st_all = (lenin_streets_strict
                     .groupby(['region', 'mun_district', 'city_type', 'city', 'shortname_2'], as_index=False)
                     .agg({'street': 'count'})
                     .rename(columns={'street': 'len_street_number'})
                    )

# выбираем переулки

cities_with_len_per = cities_with_len_st_all[(
    (cities_with_len_st_all.shortname_2 == 'переулок') |
    (cities_with_len_st_all.shortname_2 == 'проезд')    
)]

# убираем переулки из уличного фрейма

cities_with_len_st_all = cities_with_len_st_all[~(
    (cities_with_len_st_all.shortname_2 == 'переулок') |
    (cities_with_len_st_all.shortname_2 == 'проезд')    
)]

cities_with_len_per = cities_with_len_per.rename(columns={'shortname_2': 'per', 'len_street_number': 'per_number'})

cities_with_len_per.shape

need_cities = cities_with_len_per.merge(cities_with_len_st_all,
                             on=['region', 'mun_district', 'city_type', 'city',],
                             how='left'                            )

need_cities = need_cities[need_cities.len_street_number.isnull()]

need_cities.shape

need_cities

### Где любят Ленина сильнее ###

work = pd.read_csv('C:/00_Projects/Lenin/Datasets/04_work_data/work_region.csv')

weight = pd.read_csv('C:/00_Projects/Lenin/Datasets/04_work_data/work_weight.csv')

#.to_excel('C:/00_Projects/Lenin/Datasets/04_work_data/temp.xlsx', encoding='utf-8', index=False)
